refactor: Separate Parser, Converter from InputView.swift into each file
